requirements.txt
requests
robotframework
robotframework-requests
robotframework-jsonlibrary
robotframework-seleniumlibrary
# robotframework-lint
# robotframework-ride